requirements.txt
python-dotenv
requests
Flask
pylint
SQLAlchemy